refactor: web redux connect
