+++ packages/web/src/pages/sign_in.tsx
@@ -2,9 +2,46 @@ import React, { useCallback } from 'react'
 import { ExNextPageContext } from 'next'
 import { useDispatch, useSelector } from 'react-redux'
 import { signInWithGoogle } from 'src/modules/auth'
-import { SignInTemplate } from 'src/components/templates/sign_in'
 import { ReduxStore } from 'src/modules/reducer'
 import { redirect } from 'src/modules/services'
+import Paper from '@material-ui/core/Paper'
+import Typography from '@material-ui/core/Typography'
+import CssBaseline from '@material-ui/core/CssBaseline'
+import Avatar from '@material-ui/core/Avatar'
+import { makeStyles, createStyles, Theme } from '@material-ui/core/styles'
+import LockIcon from '@material-ui/icons/LockOutlined'
+import { Progress } from 'src/components/moleclues/Progress'
+import GoogleButton from 'react-google-button'
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    main: {
+      width: 'auto',
+      display: 'block', // Fix IE 11 issue.
+      marginLeft: theme.spacing(3),
+      marginRight: theme.spacing(3),
+      [theme.breakpoints.up(400 + theme.spacing(6))]: {
+        width: 400,
+        marginLeft: 'auto',
+        marginRight: 'auto'
+      }
+    },
+    paper: {
+      marginTop: theme.spacing(16),
+      display: 'flex',
+      flexDirection: 'column',
+      alignItems: 'center',
+      padding: `${theme.spacing(5)}px ${theme.spacing(3)}px ${theme.spacing(5)}px`
+    },
+    space: {
+      height: 50
+    },
+    avatar: {
+      margin: theme.spacing(2),
+      backgroundColor: theme.palette.secondary.main
+    }
+  })
+)
 
 type Props = {}
 
@@ -15,9 +52,26 @@ const SignIn = (_: Props) => {
 
   const dispatch = useDispatch()
 
-  const handleSignIn = useCallback(() => dispatch(signInWithGoogle()), [dispatch])
+  const handleSignInWithGoogle = useCallback(() => dispatch(signInWithGoogle()), [dispatch])
 
-  return <SignInTemplate isLoading={isLoading} handleSignInWithGoogle={handleSignIn} />
+  const classes = useStyles()
+
+  return (
+    <main className={classes.main}>
+      {isLoading && <Progress />}
+      <CssBaseline />
+      <Paper className={classes.paper}>
+        <Avatar className={classes.avatar}>
+          <LockIcon />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          Sign in
+        </Typography>
+        <div className={classes.space} />
+        <GoogleButton onClick={handleSignInWithGoogle} />
+      </Paper>
+    </main>
+  )
 }
 
 SignIn.getInitialProps = async (ctx: ExNextPageContext): Promise<void> => {
